Refactor weather-related classes and interfaces; update data types for temperature and wind speed, and implement OpenWeather API client methods.

+++ backend/WeatherApp/WeatherApp.Application/DTOs/Weather/WeatherDto.cs
@@ -1,7 +1,7 @@
 public class WeatherDto
 {
-    public decimal Temperature { get; set; }
-    public decimal FeelsLike { get; set; }
+    public double Temperature { get; set; }
+    public double FeelsLike { get; set; }
 
     public string Description { get; set; }
     public string Icon { get; set; }
